Clean up pipeline services (#157)
* Combine PipelineManagementBo and PipelineService

* Combine PipelineManager and PipelineService
