+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/camel/operations/detection/DetectionResponseProcessor.java
@@ -33,21 +33,21 @@ import org.mitre.mpf.wfm.camel.operations.detection.trackmerging.TrackMergingCon
 import org.mitre.mpf.wfm.data.entities.transients.*;
 import org.mitre.mpf.wfm.enums.JobStatus;
 import org.mitre.mpf.wfm.enums.MpfConstants;
-import org.mitre.mpf.wfm.pipeline.PipelineManager;
+import org.mitre.mpf.wfm.service.PipelineService;
 import org.mitre.mpf.wfm.pipeline.xml.ActionDefinition;
 import org.mitre.mpf.wfm.pipeline.xml.AlgorithmDefinition;
 import org.mitre.mpf.wfm.pipeline.xml.PropertyDefinition;
-import org.mitre.mpf.wfm.pipeline.xml.PropertyDefinitionRef;
 import org.mitre.mpf.wfm.util.AggregateJobPropertiesUtil;
 import org.mitre.mpf.wfm.util.PropertiesUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
-import java.io.File;
-import java.util.*;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.TreeMap;
 
 /** Processes the responses which have been returned from a detection component. */
 @Component(DetectionResponseProcessor.REF)
@@ -57,8 +57,7 @@ public class DetectionResponseProcessor
 	private static final Logger log = LoggerFactory.getLogger(DetectionResponseProcessor.class);
 
 	@Autowired
-	@Qualifier(PipelineManager.REF)
-	private PipelineManager pipelineManager;
+	private PipelineService pipelineService;
 
 	@Autowired
 	private PropertiesUtil propertiesUtil;
@@ -118,7 +117,7 @@ public class DetectionResponseProcessor
 
 		} else {
 			// Look for a confidence threshold.  If confidence threshold is defined, only return detections above the threshold.
-			ActionDefinition action = pipelineManager.getAction(detectionResponse.getActionName());
+			ActionDefinition action = pipelineService.getAction(detectionResponse.getActionName());
 			TransientJob job = redis.getJob(jobId, detectionResponse.getMediaId());
 
 			double confidenceThreshold = calculateConfidenceThreshold(action, job, media);
@@ -141,7 +140,7 @@ public class DetectionResponseProcessor
 				media.getMediaSpecificProperties());
 
 		if (confidenceThresholdProperty == null) {
-			AlgorithmDefinition algorithm = pipelineManager.getAlgorithm(action);
+			AlgorithmDefinition algorithm = pipelineService.getAlgorithm(action);
 			PropertyDefinition confidenceAlgorithmDef = algorithm.getProvidesCollection().getAlgorithmProperty(MpfConstants.CONFIDENCE_THRESHOLD_PROPERTY);
 			if (confidenceAlgorithmDef != null) {
 				confidenceThresholdProperty = confidenceAlgorithmDef.getDefaultValue();
