+++ trunk/workflow-manager/src/main/java/org/mitre/mpf/wfm/service/component/RemoveComponentServiceImpl.java
@@ -30,10 +30,9 @@ import org.mitre.mpf.rest.api.component.ComponentState;
 import org.mitre.mpf.rest.api.component.RegisterComponentModel;
 import org.mitre.mpf.rest.api.node.NodeManagerModel;
 import org.mitre.mpf.wfm.WfmProcessingException;
-import org.mitre.mpf.wfm.pipeline.PipelineManager;
+import org.mitre.mpf.wfm.service.PipelineService;
 import org.mitre.mpf.wfm.pipeline.xml.*;
 import org.mitre.mpf.wfm.service.NodeManagerService;
-import org.mitre.mpf.wfm.service.PipelineService;
 import org.mitre.mpf.wfm.util.PropertiesUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -53,31 +52,27 @@ public class RemoveComponentServiceImpl implements RemoveComponentService {
 
     private static final Logger _log = LoggerFactory.getLogger(RemoveComponentServiceImpl.class);
 
-    private final PipelineService pipelineService;
-
     private final NodeManagerService nodeManagerService;
 
     private final ComponentDeploymentService deployService;
 
     private final ComponentStateService componentStateService;
 
-    private final PipelineManager pipelineManager;
+    private final PipelineService pipelineService;
 
     private final PropertiesUtil propertiesUtil;
 
     @Inject
     RemoveComponentServiceImpl(
-            PipelineService pipelineService,
             NodeManagerService nodeManagerService,
             ComponentDeploymentService deployService,
             ComponentStateService componentStateService,
-            PipelineManager pipelineManager,
+            PipelineService pipelineService,
             PropertiesUtil propertiesUtil) {
-        this.pipelineService = pipelineService;
         this.nodeManagerService = nodeManagerService;
         this.deployService = deployService;
         this.componentStateService = componentStateService;
-        this.pipelineManager = pipelineManager;
+        this.pipelineService = pipelineService;
         this.propertiesUtil = propertiesUtil;
     }
 
@@ -212,41 +207,41 @@ public class RemoveComponentServiceImpl implements RemoveComponentService {
     }
 
     private void removePipeline(String pipelineName) {
-        pipelineService.removeAndDeletePipeline(pipelineName.toUpperCase());
+        pipelineService.deletePipeline(pipelineName.toUpperCase());
     }
 
     private void removeTask(String taskName) {
-        pipelineManager.getPipelines()
+        pipelineService.getPipelines()
                 .stream()
                 .filter(pd -> referencesTask(pd, taskName))
                 .map(PipelineDefinition::getName)
                 .forEach(this::removePipeline);
 
-        pipelineService.removeAndDeleteTask(taskName.toUpperCase());
+        pipelineService.deleteTask(taskName.toUpperCase());
     }
 
     private void removeAction(String actionName) {
-        pipelineManager.getTasks()
+        pipelineService.getTasks()
                 .stream()
                 .filter(td -> referencesAction(td, actionName))
                 .map(TaskDefinition::getName)
                 .forEach(this::removeTask);
 
         try {
-            pipelineService.removeAndDeleteAction(actionName.toUpperCase());
+            pipelineService.deleteAction(actionName.toUpperCase());
         } catch (WfmProcessingException e) {
             _log.error("Cannot delete action " + actionName.toUpperCase(), e);
         }
     }
 
     private void removeAlgorithm(String algorithmName) {
-        pipelineManager.getActions()
+        pipelineService.getActions()
                 .stream()
                 .filter(ad -> ad.getAlgorithmRef().equalsIgnoreCase(algorithmName))
                 .map(ActionDefinition::getName)
                 .forEach(this::removeAction);
 
-        pipelineService.removeAndDeleteAlgorithm(algorithmName.toUpperCase());
+        pipelineService.deleteAlgorithm(algorithmName.toUpperCase());
     }
 
     private void removeService(String serviceName) {
