+++ trunk/workflow-manager/src/main/resources/applicationContext-minimal.xml
@@ -35,8 +35,6 @@
 
     <context:component-scan base-package="org.mitre.mpf.wfm.util" use-default-filters="false">
         <context:include-filter type="assignable" expression="org.mitre.mpf.wfm.util.PropertiesUtil"/>
-        <context:include-filter type="assignable" expression="org.mitre.mpf.wfm.businessrules.PipelineManagementBo"/>
-        <context:include-filter type="assignable" expression="org.mitre.mpf.wfm.pipeline.PipelineManager"/>
         <context:include-filter type="assignable" expression="org.mitre.mpf.wfm.service.PipelineService"/>
         <context:include-filter type="assignable" expression="org.mitre.mpf.wfm.nodeManager.NodeManagerStatus"/>
         <context:include-filter type="assignable" expression="org.mitre.mpf.wfm.service.NodeManagerService"/>
@@ -56,8 +54,6 @@
 
     <bean class="org.mitre.mpf.wfm.util.PropertiesUtil"/>
     <bean class="org.mitre.mpf.wfm.util.IoUtils"/>
-    <bean class="org.mitre.mpf.wfm.businessrules.impl.PipelineManagementBoImpl"/>
-    <bean class="org.mitre.mpf.wfm.pipeline.PipelineManager"/>
     <bean class="org.mitre.mpf.wfm.service.PipelineServiceImpl"/>
     <bean class="org.mitre.mpf.wfm.nodeManager.NodeManagerStatus"/>
     <bean class="org.mitre.mpf.wfm.service.NodeManagerServiceImpl"/>
