+++ trunk/workflow-manager/src/test/java/org/mitre/mpf/wfm/service/component/TestAddComponentService.java
@@ -32,10 +32,11 @@ import org.junit.Before;
 import org.junit.Test;
 import org.mitre.mpf.rest.api.component.ComponentState;
 import org.mitre.mpf.rest.api.component.RegisterComponentModel;
+import org.mitre.mpf.wfm.WfmProcessingException;
 import org.mitre.mpf.wfm.enums.ActionType;
+import org.mitre.mpf.wfm.service.PipelineService;
 import org.mitre.mpf.wfm.pipeline.xml.AlgorithmDefinition;
 import org.mitre.mpf.wfm.service.NodeManagerService;
-import org.mitre.mpf.wfm.service.PipelineService;
 import org.mitre.mpf.wfm.util.Tuple;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
@@ -43,11 +44,13 @@ import org.mockito.MockitoAnnotations;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.*;
+import java.util.Collections;
+import java.util.Optional;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
-import static org.mitre.mpf.test.TestUtil.*;
+import static org.mitre.mpf.test.TestUtil.anyNonNull;
+import static org.mitre.mpf.test.TestUtil.whereArg;
 import static org.mitre.mpf.wfm.service.component.TestDescriptorConstants.*;
 import static org.mockito.Mockito.*;
 
@@ -76,9 +79,6 @@ public class TestAddComponentService {
 
     @Mock
     private RemoveComponentService _mockRemoveComponentService;
-
-    @Mock
-    private Properties _mockProperties;
 
     @Mock
     private ObjectMapper _mockObjectMapper;
@@ -178,16 +178,6 @@ public class TestAddComponentService {
         AlgorithmDefinition algoDef = new AlgorithmDefinition(
                 ActionType.DETECTION, descriptor.algorithm.name.toUpperCase(), descriptor.algorithm.description);
 
-        when(_mockPipelineService.addAndSaveAlgorithm(algoDef))
-                .thenReturn(_successTuple);
-
-        when(_mockPipelineService.addAndSaveActionDeprecated(
-					contains(algoDef.getName()), anyString(), eq(algoDef.getName()), eq(Collections.emptyMap())))
-                .thenReturn(_successTuple);
-
-        when(_mockPipelineService.addAndSaveTaskDeprecated(whereArg(td -> td.getName().contains(algoDef.getName()))))
-                .thenReturn(_successTuple);
-
         when(_mockStateService.getByPackageFile(_testPackageName))
                 .thenReturn(Optional.of(registerModel));
 
@@ -221,11 +211,15 @@ public class TestAddComponentService {
 
         // Verify mocked methods
         verify(_mockPipelineService)
-                .addAndSaveAlgorithm(algoDef);
+                .saveAlgorithm(algoDef);
+
+	    verify(_mockPipelineService)
+                .saveAction(whereArg(ad -> ad.getName().contains(algoDef.getName())
+                        && ad.getAlgorithmRef().equals(algoDef.getName())
+                        && ad.getProperties().isEmpty() ));
 
         verify(_mockPipelineService)
-                .addAndSaveActionDeprecated(contains(algoDef.getName()), anyString(), anyString(),
-                                            eq(Collections.emptyMap()));
+                .saveTask(whereArg(td -> td.getName().contains(algoDef.getName())));
 
         verify(_mockDeploymentService)
                 .deployComponent(_testPackageName);
@@ -284,19 +278,6 @@ public class TestAddComponentService {
         AlgorithmDefinition algoDef = new AlgorithmDefinition(
                 ActionType.DETECTION, descriptor.algorithm.name.toUpperCase(), descriptor.algorithm.description);
 
-        when(_mockPipelineService.addAndSaveAlgorithm(algoDef))
-                .thenReturn(_successTuple);
-
-        when(_mockPipelineService.addAndSaveActionDeprecated(nonBlank(), nonBlank(), eq(REFERENCED_ALGO_NAME),
-                anyNonNull()))
-                .thenReturn(_successTuple);
-
-        when(_mockPipelineService.addAndSaveTaskDeprecated(anyNonNull()))
-                .thenReturn(_successTuple);
-
-        when(_mockPipelineService.addAndSavePipelineDeprecated(anyNonNull()))
-                .thenReturn(_successTuple);
-
         when(_mockNodeManager.getServiceModels())
                 .thenReturn(Collections.singletonMap("fake name", null));
 
@@ -316,31 +297,32 @@ public class TestAddComponentService {
                                 && rcm.getTasks().containsAll(TASK_NAMES)
                                 && rcm.getPipelines().contains(PIPELINE_NAME)));
 
+        verify(_mockPipelineService)
+                .saveAlgorithm(algoDef);
+
         verify(_mockPipelineService, times(3))
-                .addAndSaveActionDeprecated(nonBlank(), nonBlank(), eq(REFERENCED_ALGO_NAME), anyNonNull());
+                .saveAction(whereArg(ad -> ad.getAlgorithmRef().equals(REFERENCED_ALGO_NAME)));
 
         verify(_mockPipelineService)
-                .addAndSaveActionDeprecated(
-                        eq(ACTION_NAMES.get(0)),
-                        nonBlank(),
-                        nonBlank(),
-                        whereArg(m -> m.get(ACTION1_PROP_NAMES.get(0))
-                                .equals(ACTION1_PROP_VALUES.get(0))));
+                .saveAction(whereArg(ad -> ad.getName().equals(ACTION_NAMES.get(0))
+                        && ad.getProperties().stream()
+                                .anyMatch(pd -> pd.getName().equals(ACTION1_PROP_NAMES.get(0))
+                                        && pd.getValue().equals(ACTION1_PROP_VALUES.get(0)))));
 
         verify(_mockPipelineService)
-                .addAndSaveTaskDeprecated(whereArg(t ->
+                .saveTask(whereArg(t ->
                         t.getName().equals(TASK_NAMES.get(0))
                                 && t.getDescription().equals(TASK_NAMES.get(0) + " description")
                                 && t.getActions().size() == 1));
 
         verify(_mockPipelineService)
-                .addAndSaveTaskDeprecated(whereArg(t ->
+                .saveTask(whereArg(t ->
                         t.getName().equals(TASK_NAMES.get(1))
                                 && t.getDescription().equals(TASK_NAMES.get(1) + " description")
                                 && t.getActions().size() == 2));
 
         verify(_mockPipelineService)
-                .addAndSavePipelineDeprecated(whereArg(p ->
+                .savePipeline(whereArg(p ->
                         p.getName().equals(PIPELINE_NAME)
                                 && p.getDescription().contains("description")
                                 && p.getTaskRefs().size() == 2));
@@ -358,8 +340,8 @@ public class TestAddComponentService {
 
         setUpMocksForDescriptor(descriptor);
 
-        when(_mockPipelineService.addAndSaveAlgorithm(any()))
-                .thenReturn(_failureTuple);
+        doThrow(WfmProcessingException.class)
+                .when(_mockPipelineService).saveAlgorithm(any());
 
         // Act
         try {
@@ -391,10 +373,10 @@ public class TestAddComponentService {
         }
 
         verify(_mockPipelineService, never())
-                .addAndSaveAlgorithm(any());
+                .saveAlgorithm(any());
 
         verify(_mockPipelineService, never())
-                .addAndSaveActionDeprecated(any(), any(), any(), any());
+                .saveAction(any());
         assertUndeployed(COMPONENT_NAME);
     }
 
@@ -414,10 +396,10 @@ public class TestAddComponentService {
         }
 
         verify(_mockPipelineService, never())
-                .addAndSaveAlgorithm(any());
+                .saveAlgorithm(any());
 
         verify(_mockPipelineService, never())
-                .addAndSaveActionDeprecated(any(), any(), any(), any());
+                .saveAction(any());
         assertUndeployed(COMPONENT_NAME);
     }
