+++ trunk/workflow-manager/src/test/java/org/mitre/mpf/wfm/service/component/TestCustomPipelineValidator.java
@@ -26,16 +26,22 @@
 
 package org.mitre.mpf.wfm.service.component;
 
+import com.google.common.collect.ImmutableSortedSet;
 import org.junit.Before;
 import org.junit.Test;
+import org.mitre.mpf.wfm.enums.ActionType;
+import org.mitre.mpf.wfm.service.PipelineService;
+import org.mitre.mpf.wfm.pipeline.xml.AlgorithmDefinition;
 import org.mitre.mpf.wfm.pipeline.xml.PropertyDefinition;
 import org.mitre.mpf.wfm.pipeline.xml.ValueType;
-import org.mitre.mpf.wfm.service.PipelineService;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
-import java.util.*;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 import static java.util.stream.Collectors.toList;
 import static org.junit.Assert.*;
@@ -65,13 +71,13 @@ public class TestCustomPipelineValidator {
         descriptor.pipelines.get(0).tasks.add(existingTaskRef);
 
         when(_mockPipelineService.getAlgorithmNames())
-                .thenReturn(Arrays.asList("foo", REFERENCED_ALGO_NAME));
+                .thenReturn(ImmutableSortedSet.of("foo", REFERENCED_ALGO_NAME));
+
 
         when(_mockPipelineService.getActionNames())
-                .thenReturn(Arrays.asList("bar", existingActionRef));
-
+                .thenReturn(ImmutableSortedSet.of("bar", existingActionRef));
         when(_mockPipelineService.getTaskNames())
-                .thenReturn(Collections.singletonList(existingTaskRef));
+                .thenReturn(ImmutableSortedSet.of(existingTaskRef));
 
 
         List<PropertyDefinition> propDefs = ACTION1_PROP_NAMES
@@ -80,11 +86,11 @@ public class TestCustomPipelineValidator {
                 .collect(toList());
         propDefs.add(new PropertyDefinition("foo", ValueType.INT, "2", "0"));
 
-        when(_mockPipelineService.getAlgorithmProperties(REFERENCED_ALGO_NAME))
-                .thenReturn(propDefs);
+        setupAlgoWithProperties(REFERENCED_ALGO_NAME, propDefs);
 
         _pipelineValidator.validate(descriptor);
     }
+
 
 
     @Test
@@ -92,13 +98,13 @@ public class TestCustomPipelineValidator {
         JsonComponentDescriptor descriptor = TestDescriptorFactory.getWithCustomPipeline();
 
         when(_mockPipelineService.getActionNames())
-                .thenReturn(ACTION_NAMES.subList(0, 2));
+                .thenReturn(ImmutableSortedSet.copyOf(ACTION_NAMES.subList(0, 2)));
 
         when(_mockPipelineService.getTaskNames())
-                .thenReturn(Arrays.asList("foo", TASK_NAMES.get(0)));
+                .thenReturn(ImmutableSortedSet.of("foo", TASK_NAMES.get(0)));
 
         when(_mockPipelineService.getPipelineNames())
-                .thenReturn(Collections.singletonList(PIPELINE_NAME));
+                .thenReturn(ImmutableSortedSet.of(PIPELINE_NAME));
 
         try {
             _pipelineValidator.validate(descriptor);
@@ -128,7 +134,7 @@ public class TestCustomPipelineValidator {
         descriptor.pipelines.get(0).tasks.add(invalidTaskName);
 
         when(_mockPipelineService.getAlgorithmNames())
-                .thenReturn(Arrays.asList("foo", "bar", "baz"));
+                .thenReturn(ImmutableSortedSet.of("foo", "bar", "baz"));
 
 
         try {
@@ -158,8 +164,7 @@ public class TestCustomPipelineValidator {
                 .stream()
                 .map(n -> new PropertyDefinition(n, ValueType.INT, "1", "0"))
                 .collect(toList());
-        when(_mockPipelineService.getAlgorithmProperties(REFERENCED_ALGO_NAME))
-                .thenReturn(propDefs);
+        setupAlgoWithProperties(REFERENCED_ALGO_NAME, propDefs);
 
 
         try {
@@ -195,8 +200,7 @@ public class TestCustomPipelineValidator {
                 .stream()
                 .map(n -> new PropertyDefinition(n, ValueType.INT, "1", "0"))
                 .collect(toList());
-        when(_mockPipelineService.getAlgorithmProperties(REFERENCED_ALGO_NAME))
-                .thenReturn(propDefs);
+        setupAlgoWithProperties(REFERENCED_ALGO_NAME, propDefs);
 
         try {
             _pipelineValidator.validate(descriptor);
@@ -216,5 +220,14 @@ public class TestCustomPipelineValidator {
     @Test
     public void descriptorWithoutCustomPipelinesValidates() throws InvalidCustomPipelinesException {
         _pipelineValidator.validate(TestDescriptorFactory.get());
+    }
+
+
+    private void setupAlgoWithProperties(String algoName, List<PropertyDefinition> properties) {
+        AlgorithmDefinition algoDef = new AlgorithmDefinition(ActionType.DETECTION, algoName, "");
+        algoDef.getProvidesCollection().getAlgorithmProperties().addAll(properties);
+
+        when(_mockPipelineService.getAlgorithm(algoName))
+                .thenReturn(algoDef);
     }
 }
